Pushed reusable code up the class hierarchy (#186)
