+++ dz3-swing/src/main/java/net/sf/dz3/view/swing/sensor/SensorChart.java
@@ -1,13 +1,36 @@
 package net.sf.dz3.view.swing.sensor;
 
 import com.homeclimatecontrol.jukebox.datastream.signal.model.DataSample;
-import com.homeclimatecontrol.jukebox.datastream.signal.model.DataSink;
+import net.sf.dz3.view.swing.AbstractChart;
 
-import javax.swing.JPanel;
+import java.awt.Dimension;
+import java.awt.Graphics2D;
+import java.awt.Insets;
+import java.time.Clock;
 
-public class SensorChart extends JPanel implements DataSink<Double> {
+public class SensorChart extends AbstractChart<Double> {
+
+    protected SensorChart(Clock clock, long chartLengthMillis) {
+        super(clock, chartLengthMillis);
+    }
+
     @Override
     public void consume(DataSample<Double> signal) {
+
+    }
+
+    @Override
+    protected boolean isDataAvailable() {
+        return false;
+    }
+
+    @Override
+    protected Limits recalculateVerticalLimits() {
+        return null;
+    }
+
+    @Override
+    protected void paintCharts(Graphics2D g2d, Dimension boundary, Insets insets, long now, double xScale, long xOffset, double yScale, double yOffset) {
 
     }
 }
